requirements.txt
Flask==0.9
gunicorn==0.17.2
requests==2.4.3
Flask-SQLAlchemy==2.0
psycopg2==2.5.4